# docs/swagger.yaml
basePath: /
definitions:
...
  controllers.ReqIssueTokenForDareid:
    properties:
      dareid:
        type: string
      data:
        type: string
    required:
    - dareid
    type: object
...
info:
  contact:
    url: https://github.com/0xbarchitect
  description: All-in-one SSO Service
  license:
    name: Apache 2.0
    url: http://www.apache.org/licenses/LICENSE-2.0.html
  title: SSO
  version: "1.0"
paths:
...
  /login:
    get:
      operationId: LoginHandler
      parameters:
      - description: state
        in: formData
        name: state
        type: string
      - description: username
        in: formData
        name: username
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: wallet_address
        in: formData
        name: wallet_address
        type: string
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: LoginHandler
    post:
      operationId: LoginHandler
      parameters:
      - description: state
        in: formData
        name: state
        type: string
      - description: username
        in: formData
        name: username
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: wallet_address
        in: formData
        name: wallet_address
        type: string
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: LoginHandler
...
  /oauth/token:
    post:
      consumes:
      - application/x-www-form-urlencoded
      operationId: OauthTokenHandler
      parameters:
      - description: 'Grant type, eg.: client_credentials, password, authorization_code,
          refresh_token'
        in: formData
        name: grant_type
        required: true
        type: string
      - description: authorization code
        in: formData
        name: code
        type: string
      - description: client's redirect uri
        in: formData
        name: redirect_uri
        type: string
      - description: PKCE code verifier
        in: formData
        name: code_verifier
        type: string
      - description: refresh token
        in: formData
        name: refresh_token
        type: string
      - description: scope of refreshing token
        in: formData
        name: scope
        type: string
      - description: 'password gt: username'
        in: formData
        name: username
        type: string
      - description: 'password gt: password'
        in: formData
        name: password
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ClientBasicAuth: []
      summary: OauthTokenHandler
  /oauth/token-with-data:
    post:
      deprecated: true
      operationId: IssueTokenForDareidWithData
      parameters:
      - description: issue token for dareid req body
        in: body
        name: _
        schema:
          $ref: '#/definitions/controllers.ReqIssueTokenForDareid'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - ClientBasicAuth: []
      summary: IssueTokenForDareidWithData - this API is deprecated, please use /oauth/token
        instead
  /oauth/validate-token:
    get:
      operationId: ValidateToken
      parameters:
      - description: oauth token
        in: query
        name: access_token
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            type: string
      summary: ValidateToken
...
  /signup:
    get:
      operationId: SignupHandler
      parameters:
      - description: state
        in: formData
        name: state
        type: string
      - description: username
        in: formData
        name: username
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: confirm
        in: formData
        name: confirm
        type: string
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JwtHeader: []
      summary: SignupHandler
    post:
      operationId: SignupHandler
      parameters:
      - description: state
        in: formData
        name: state
        type: string
      - description: username
        in: formData
        name: username
        type: string
      - description: password
        in: formData
        name: password
        type: string
      - description: confirm
        in: formData
        name: confirm
        type: string
      produces:
      - multipart/form-data
      responses:
        "200":
          description: OK
          schema:
            type: string
      security:
      - JwtHeader: []
      summary: SignupHandler
...
securityDefinitions:
  ClientBasicAuth:
    type: basic
  JwtHeader:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"